[Feat] Added some additional error handling

+++ src/main/java/weijie/duke/commands/DeleteCommand.java
@@ -1,5 +1,6 @@
 package weijie.duke.commands;
 
+import weijie.duke.exceptions.DukeInvalidInputException;
 import weijie.duke.models.Task;
 import weijie.duke.repos.IRepository;
 import weijie.duke.responses.TaskResponse;
@@ -14,8 +15,20 @@ public class DeleteCommand implements ITaskCommand {
     }
 
     @Override
+    @SuppressWarnings("Duplicates")
     public TaskResponse execute(String... args) {
+        if (args.length < 2) {
+            return new TaskResponse(
+                    new DukeInvalidInputException("☹ OOPS!!! Please input the number of the task to mark as done."));
+        }
+
         int id = Integer.parseInt(args[1]) - 1;
+
+        if (id > repo.getSize()) {
+            return new TaskResponse(
+                    new DukeInvalidInputException("☹ OOPS!!! Task with that number does not exist!"));
+        }
+
         Task deletedTask = repo.get(id);
         repo.delete(id);
         int size = repo.getSize();
